+++ application/controllers/DossierController.php
@@ -207,6 +207,8 @@ class DossierController extends Zend_Controller_Action
 
     public function indexAction()
     {
+        $this->view->headScript()->appendFile('/js/tinymce.min.js');
+
         $this->view->do = "new";
         if ($this->_getParam("id")) {
             $this->view->do = "edit";
@@ -1190,10 +1192,43 @@ class DossierController extends Zend_Controller_Action
                     $listeEtab = $DBetablissementDossier->getEtablissementListe($idDossier);
                 }
 
+                $cache = Zend_Controller_Front::getInstance()->getParam('bootstrap')->getResource('cache');
+                $mygroupe = Zend_Auth::getInstance()->getIdentity()['group']['LIBELLE_GROUPE'];
+                $service_ets = new Service_Etablissement;
+                $arrayEtsAvis = array();
+                if (unserialize($cache->load('acl'))->isAllowed($mygroupe, "alerte_email", "alerte_avis") 
+                    && getenv('PREVARISC_MAIL_ENABLED') && getenv('PREVARISC_MAIL_ENABLED') == 1) {
+                    foreach($listeEtab as $etab) {
+                        $currentEts = $service_ets->get($etab['ID_ETABLISSEMENT']);
+                        if ($currentEts) {
+                            $arrayEtsAvis[$etab['ID_ETABLISSEMENT']]['avis'] = $currentEts['avis'];
+                            $arrayEtsAvis[$etab['ID_ETABLISSEMENT']]['libelle'] = 
+                                $currentEts['informations']['LIBELLE_ETABLISSEMENTINFORMATIONS'];
+                        }
+                    }
+                }
+
                 $dbEtab = new Model_DbTable_Etablissement();
 
                 $updatedEtab = $service_dossier->saveDossierDonnantAvis($nouveauDossier, $listeEtab, $cache, $this->_getParam('repercuterAvis'));
-
+                if (unserialize($cache->load('acl'))->isAllowed($mygroupe, "alerte_email", "alerte_avis")
+                    && getenv('PREVARISC_MAIL_ENABLED') && getenv('PREVARISC_MAIL_ENABLED') == 1) {
+                    $service_alerte = new Service_Alerte;
+                    foreach($updatedEtab as $upEts) {
+                        if ($upEts['ID_DOSSIER_DONNANT_AVIS'] === $nouveauDossier['ID_DOSSIER']
+                            && $nouveauDossier['AVIS_DOSSIER_COMMISSION'] !== 
+                                $arrayEtsAvis[$upEts['ID_ETABLISSEMENT']]['avis']) {
+                            $options = $service_alerte->getLink(2, $upEts["ID_ETABLISSEMENT"]);    
+                            $this->_helper->flashMessenger(array(
+                            'context' => 'success',
+                            'title' => 'Mise à jour réussie !', 'message' => 'L\'établissement '
+                            . $arrayEtsAvis[$upEts['ID_ETABLISSEMENT']]['libelle'] 
+                            . ' a bien été mis à jour.' . $options));    
+                        }
+                        
+                    }
+                }
+                
                 // AVERTISSEMENT SUR L'OUVERTURE D'UN ETABLISSEMENT A EFFECTUER
                 // Dans le cas d'une visite avant ouverture avec avis de commission positif
                 if ($this->_getParam("AVIS_DOSSIER_COMMISSION") == 1 && in_array($idNature, array(47, 48)))