+++ application/controllers/EtablissementController.php
@@ -5,6 +5,7 @@ class EtablissementController extends Zend_Controller_Action
     public function indexAction()
     {
         $this->_helper->layout->setLayout('etablissement');
+        $this->view->headScript()->appendFile('/js/tinymce.min.js');
 
         $service_etablissement = new Service_Etablissement;
         $service_groupement_communes = new Service_GroupementCommunes;
@@ -76,9 +77,21 @@ class EtablissementController extends Zend_Controller_Action
         if($this->_request->isPost()) {
             try {
                 $post = $this->_request->getPost();
+                $options = '';
+                if (getenv('PREVARISC_MAIL_ENABLED') && getenv('PREVARISC_MAIL_ENABLED') == 1) {
+                    $typeAlerte = $service_etablissement->checkAlerte($etablissement, $post);
+
+                    if (unserialize($cache->load('acl'))->isAllowed($mygroupe, "alerte_email", "alerte_statut", "alerte_classement")) {
+                        if ($typeAlerte !== false) {
+                            $service_alerte = new Service_Alerte;
+                            $options = $service_alerte->getLink($typeAlerte);
+                        }
+                    }    
+                }
+                
                 $date = date("Y-m-d");
                 $service_etablissement->save($post['ID_GENRE'], $post, $this->_request->id, $date);
-                $this->_helper->flashMessenger(array('context' => 'success', 'title' => 'Mise à jour réussie !', 'message' => 'L\'établissement a bien été mis à jour.'));
+                $this->_helper->flashMessenger(array('context' => 'success', 'title' => 'Mise à jour réussie !', 'message' => 'L\'établissement a bien été mis à jour.' . $options));
                 $this->_helper->redirector('index', null, null, array('id' => $this->_request->id));
             }
             catch(Exception $e) {